pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-langgraph-quickstart"
version = "0.1.0"
description = "LangGraph agent with multi-server MCP tool support on Azure"
requires-python = ">=3.11"

dependencies = [
    # LangChain / LangGraph
    "langgraph>=0.2",
    "langchain-core>=0.2",
    "langchain-openai>=0.1",

    # Azure OpenAI (also provided via langchain-openai)
    "openai>=1.30",

    # Web framework
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",

    # Config / HTTP
    "pyyaml>=6.0",
    "httpx>=0.27",

    # Type helpers
    "typing-extensions>=4.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx",          # for TestClient
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
